Move TechnologyIncomplete to components
