+++ packages/frontend/src/stories/components/project/TechnologyIncomplete.stories.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import { TechnologyIncomplete as TechnologyIncompleteComponent } from '../../../components/project/TechnologyIncomplete'
+
+export default {
+  title: 'Components/Project/TechnologyIncomplete',
+}
+
+export function TechnologyIncomplete() {
+  return (
+    <div className="leading-normal p-4">
+      <TechnologyIncompleteComponent editLink="#" twitterLink="#" />
+    </div>
+  )
+}
